Move Haskell API to version package
We don't want `purty` to be responsible for templating.

The version templating stuff isn't really source code that we want to
think about when building the `purty` tool. We just want the version to
exist somehow, somewhere. We move the code over to that location, so we
don't have to think about it.

We also take this time to start naming things for the future where we
build more than one tool. Ideally, this repo will become the place for
all these PureScript tools. So, we give it the name `purs-tool`. Might
require a change later on.