Move rest of `lib` to `src` package
All of this stuff is for the real program part of things.

We've removed everything that had a slightly different concern from
formatting. The only stuff left had to do with running the CLI. We move
it all to the same package so we can keep all related code in line.