Lock down visibility
We don't want a web of dependencies.

The idea is that certain packages should only be consumed by others, not
that every package should be able to consume any other package.

`pkg` is the furthest from our core problem we're solving. It's stuff
that might leave the repo. It's hard to let it leave the repo if it is
coupled to stuff we're doing internally or in our programs.

`internal` is stuff we're doing that is either cross-cutting or core to
the problem (but not directly the program). `pkg` shouldn't be able to
depend on it, but our programs will have to.

`cmd` is the actual programs we're producing. They should definitely not
be depended on by `internal` or `pkg` because they have no utility
outside of themselves. We still need to build and deploy them though, so
we have other packages able to depend on this stuff. The fact that tests
specific to the program aren't with the program is a bit jarring. We
should see about combining these packages.

Lastly, we drop the visibility of `//ci/travis:purty-tar`. It's only
used internal to that package. We've no need to make it visible to every
subpackage in the workspace.