Split out nested cases
Each value should do one thing.

This is taking a page from other languages where having a bunch of
nesting or performing multiple things in a single value is considered a
smell. In the Haskell ecosystem, this mindset isn't as prevalent. So,
seeing multiple cases is fairly common. After working with deeply nested
case stuff for a while, it's become clearer that that's not the way
forward.

Yes, the type system makes sure everything is there. But sometimes, type
inference doesn't work out the way we expect. There's times where the
right thing to do is pull up a value to the top-level and give it a
type. Inference works better, the arguments to it are clearer, and the
nesting goes away.