Move module formatting to its own package
We want to push more of the format logic to `//interal/format`.

The `Purty` module was doing quite a bit of work for stuff it didn't
have to. It shouldn't have to annotate and format things. Annotation is
largely an implementation detail of the way the `//internal/format` is
implemented. If we don't force external consumers to know about
annotation, we can not expose that module and have an easier API to deal
with.

We move implementation details for formatting modules, to the
`//internal/format` package. We'll probably want to split based on
syntactic constructs, but we'll see how far we get.