Inline `Purty` module to `Args`
The `Purty` module wasn't really pulling its weight anymore.

For a long time, it's been questionable whether having both `Args` and
`Purty` was a good idea. We exported stuff from `Args` specifically for
`Purty` to import. But that meant that whenever something changed in
`Args`, we'd have to update it in `Purty` as well.

We started moving stuff out of `Purty` and into their respective
modules. Since then, `Purty` has been dwindling. It got to the point
that it didn't really make a whole lot of sense to keep it around.

We inline the whole module into `Args`, and let it deal with its
internals instead of forcing it to expose a bunch of stuff it doesn't
need to. It's nice that we no long have to export so much in order for
things to work.