Be explicit about AppVeyor
We finally hit the timeout when building everything on AppVeyor.

We've been taking the lazy approach and telling everything to build on
AppVeyor. This works well because it means we don't have to keep
anything lined up inside the bazel build. It doesn't work so well when
we spend nearly an hour building `ormolu` and its dependencies.

We try only building what we need and bailing on anything else.