Move `alex` dependency to workspace
Other dependencies also need `alex` to build correctly.

Some of our other dependencies (like `purescript-cst`) aslo need the
`alex` dependency to be available in order to build correctly.

We were only incidentally able to compile things properly. Since
`ormolu` had to be built for things to build, we were always getting the
`alex` dependency. That means that if we were to dump all of the
`ormolu` dependencies, the rest of the code base wouldn't build anymore.

Instead of having the dependencies be incorrect like that, we push the
`alex` dependency up to the workspace-level so it's more like the
`happy` dependency.