Move `purescript-cst` to a `bazel` rule
We want to work around an issue with packages not at the top-level.

It looks as though there's some issue in the newer versions of
`rules_haskell` for Haskell packages that aren't at the top-level. This
comment gives an overview of the problem:
https://github.com/tweag/rules_haskell/issues/1392#issuecomment-657472777.
To paraphrase, Haskell packages in a sub-directory don't really work
with `stack ls dependencies json` at the moment.

Since we want to update `rules_haskell` to a newer version, we need to
address this. We can make this change now without updating the version
so that the version bump of `rules_haskell` has a smaller diff.

In all likelihood, this issue will be fixed at some point. Haskell
packages in a sub-directory are fairly common. Until that day, we follow
most of the workaround laid out in that GitHub issue: pull in the
package as a `bazel` `http_archive` rule, point `stack_snapshot` to the
package as a `vendored_package`.

The part we do a bit differently is not stripping down to the
sub-directory. Instead, we treat the whole archive as one thing and
locate the files based on where they live from the top-level. We take
this approach primarily because we're going to pull in both
`purescript-ast` and `purescript-cst`.

Since the commit we're using doesn't have a
`lib/purescript-cst/purescript-cst.cabal` file, we have to patch it to
add that.