Pin stackage dependencies
This should address some of the caching issues we run into.

As mentioned in this GitHub issue:
https://github.com/tweag/rules_haskell/issues/1293, there are issues
with `stack` running intermittently locally, and with caching not really
working on any of the CI systems we use. It looks like the
`stack_snapshot_json` argument should help with that.

After working with it for a little while, it looks like we consistently
get sub-second `bazel test //...` when it should be a no-op! This is the
behavior we wanted, and it's here!

We'll have to see how this plays out on CI.

We also add a rule in the `Makefile` to help with regenerating that
pinned file when we make changes.