Patch `rules_haskell` to ignore `unix` on Windows
We don't have a `unix` toolchain package when we're on Windows.

Something is different between the last version of `rules_haskell` we
were using: b9db31a285e697f98a2a9d4afebefb2aacc6b05a and v0.13. It's
unclear what the difference is, but in v0.13 the Windows build seems to
want to make a `haskell_toolchain_library` rule for the `unix` package.
The `unix` package doesn't exist on GHC for Windows. So we cannot use
it.

We try and unblock ourselves by patching `rules_haskell`. If this works,
we should raise the issue upstream to see if there's maybe something we
are forgetting to do or just doing incorrectly. It could be the case
that we've always been doing something incorrectly and it was only
accidentally working up until this point. In any case, we can cross that
bridge when we come to it.