Use an older image to address `unix` on Windows
We try using an older image that is closer to what upstream uses.

In the `rules_haskell` repo, they use an image named `vs2017-win2016`:
https://github.com/tweag/rules_haskell/blob/c858a1d6c8d9866794762c6ca782841f55710907/azure-pipelines.yml#L4.
The pre-installed software for that image is listed here:
https://github.com/actions/virtual-environments/blob/f8876950750757dc5ee322023c4b99246c6f438c/images/win/Windows2016-Readme.md.
The pre-installed software for AppVeyor is listed here:
https://www.appveyor.com/docs/windows-images-software/.

At first glance, they look compatible. We'll have to see if this works
or not by pushing it up.