Add a `stack init`
Looks like there's nothing initialized for this stack.

We're cargo culting the suggestion in the issue that we raised:
https://github.com/tweag/rules_haskell/issues/1443#issuecomment-738815933.
To that end, it's unclear if it's expected that there would be an error
about `stack init` or if it was forgotten in the suggestion.

Given the further explanation that the idea is to understand where the
`unix` dependency is coming from:
https://github.com/tweag/rules_haskell/issues/1443#issuecomment-740493577,
it seems reasonable that we want the `stack init` here.

Although, given that we don't really know how `rules_haskell` works and
what the steps in the process are, we might have discovered that the
`unix` dependency doesn't come from `stack` but from something else.

In any case, we add the `stack init` to the fray preemptively just so we
have more information in case it's needed later on.