Try without `stack_snapshot_json`
It seems like this might be causing Windows to pick up `unix`.

As mentioned in
https://github.com/tweag/rules_haskell/issues/1443#issuecomment-762745849,
the `stack_snapshot_json` attempts to load all dependencies for all
platforms. Since that lock file has the `unix` dependency (for Linux and
macOS), it seem that Windows is also picking that up.

We try removing that. Hopefully, this will get us further along.