Use an older image to circumvent .NET 5.0
We try using an older image that is closer to what upstream uses.

In the `rules_haskell` repo, they use an image named `vs2017-win2016`:
https://github.com/tweag/rules_haskell/blob/c858a1d6c8d9866794762c6ca782841f55710907/azure-pipelines.yml#L4.
The pre-installed software for that image is listed here:
https://github.com/actions/virtual-environments/blob/f8876950750757dc5ee322023c4b99246c6f438c/images/win/Windows2016-Readme.md.
The pre-installed software for AppVeyor is listed here:
https://www.appveyor.com/docs/windows-images-software/.
It seems like the `Visual Studio 2017` image doesn't have .NET 5.0 and
its changes to how the C toolchain work.