Add golden tests for new syntax
Kind signatures and roles were added, so validate how we format them.

For kind signatures, we follow what we do for type signatures:
- Don't put a newline after the signature.
- Format over multiple lines if there's a break.

For roles, we take an approach similar to fixities:
- Put everything on a single line no matter what.

These definitely aren't the only way to format these new constructs, but
it should be non-controversial to format them this way.