Bump `purescript-ast`/`purescript-cst` to 0.14.0
We upgrade our dependencies to the latest versions.

This change is fairly straight-forward. We're mostly just updating any
of the stuff that worked with kinds to now work with types.

We also make sure we format according to our golden tests. The change
for the now deprecated octothorpe row constructor was basically moved
whole-cloth.

What's nice is that it seems like we can still parse valid 0.13.x syntax
as well. Which means this ought not be a breaking change.