Remove path to GNU binaries
We no longer need to put the GNU binaries on the path for macOS.

The only reason we had this path was for macOS to have access to the GNU
binaries so we didn't have to write the bash scripts a certain way. Now
that we're relying on the POSIX behavior, we ought to be able to use
whatever binaries are on the path (since macOS has POSIX-compliant
binaries).